Added support for horizontal lines, such as a y-axis or value thresholds.
`chart.options.horizontalLines` is an array of values of form:

    { color:'#ffffff', lineWidth: 1, value: 0 }

That example results in a 1px white line, indicating the y-axis. This is useful if your chart has both positive and negative values, and absolute values are important.  Before this patch, only relative values could easily be observed.

The set of lines may be dynamically updated at runtime. The example in customisation.html demonstrates this.