Old browsers with multiple charts in document don't flicker (#40)
The shared 'lastTime' variable in the fallback function for requestAnimationFrame caused thrashing and inconsistent spacing of frames when more than one animation loop was run. Instead, frames are now requested for 16 ms away, which in the ideal case gives 60 fps. In practice it is slightly lower, however it is still nice and smooooooth.

You can test this by commenting out all the lines like: window.requestAnimationFrame, etc.

Thanks to Bai Yang (@asbai) for providing the logic for this patch.